added abs on false(sz)
use of false() as a means of getting an array of zeros on sz resulted in an empty dimension when a negative is passed in.  The previous structure seemed to be predicated on a presumption of either RAS or LPI.  Use of Abs(sz) ensures that the actual dimensions are passed.